Simplify Library and separating
