--- app/library/mistral.py
@@ -1,73 +1,0 @@
-from pkg.mistral import MistralLLM, ChatMistralAI
-from pkg.vectorstore.chromadb import VectorStoreRetriever
-from app.library.vectorstore import Vectorstores
-from app import redis
-from pkg.chain import Chain, Runnable
-from pkg.runnable import RunnableChain, RunnableWithMessageHistory
-from pkg.history import RedisChatMessageHistory
-from pkg.chain.prompter import PromptTemplate
-
-class MistralAILibrary(object):
-    def __init__(self, vectorstores: Vectorstores, llm: MistralLLM, model: str, redis: redis):
-        self.vectorstore = vectorstores
-        self.mistral = llm
-        self.redis = redis
-        self.model = model
-        self.chain = Chain()
-        self.runnable = RunnableChain()
-    
-    def retriever(self, vector, top_k, fetch_k, collection) -> VectorStoreRetriever:
-        if top_k is None:
-            top_k = 3
-        if fetch_k is None:
-            fetch_k = 10
-        try:
-            vectordb = self.vectorstore.configure(vector)
-            return vectordb.retriever(
-                topK=top_k,
-                fetchK=fetch_k,
-                collection=collection
-            )
-        except Exception as e:
-            raise e
-
-    
-    def get_llm(self, model) -> ChatMistralAI:
-        try:
-            return self.mistral.run(
-                redis_url=self.redis.str_conn(),
-                model=model
-            )
-        except Exception as e:
-            raise e
-
-    def retrieval(self, promp_tpl: PromptTemplate, retriever: VectorStoreRetriever) -> Runnable:
-        llm = self.get_llm(self.model)
-        try:
-            return self.chain.retrieval(
-                prompt_template=promp_tpl,
-                platform=llm,
-                retriever=retriever
-            )
-        except Exception as e:
-            raise e
-
-    def chain_with_history(
-            self, 
-            retrival: Runnable, 
-            history: RedisChatMessageHistory, 
-            input_messages_key: str,
-            history_messages_key: str,
-            output_messages_key: str
-        ) -> RunnableWithMessageHistory:
-        
-        try:
-            return self.runnable.chain_with_history(
-                retrieval=retrival,
-                history=history,
-                input_messages_key=input_messages_key,
-                history_messages_key=history_messages_key,
-                output_messages_key=output_messages_key,
-            )
-        except Exception as e:
-            raise e
