--- pkg/chain/history.py
@@ -1,16 +1,0 @@
-from langchain_core.runnables.history import RunnableWithMessageHistory
-from langchain_community.chat_message_histories import RedisChatMessageHistory
-from langchain.chains.retrieval import Runnable
-
-def chain_with_history(retrieval: Runnable,
-            history: RedisChatMessageHistory,
-            input_messages_key,
-            history_messages_key,
-            output_messages_key) -> RunnableWithMessageHistory:
-        return RunnableWithMessageHistory(
-            retrieval,
-            lambda session_id: history,
-            input_messages_key=input_messages_key,
-            history_messages_key=history_messages_key,
-            output_messages_key=output_messages_key
-        )
