+++ pkg/vectorstore/__init__.py
@@ -1,0 +1,1 @@
+from langchain_core.vectorstores import VectorStoreRetriever
